Fix: Fix api endpoint at command magician
